pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "llm-router"
version = "0.1.0"
description = "Lightweight LLM API router with priority cascading, rate limit tracking, and model scouting."
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
authors = [
    {name = "donjguido"},
]
keywords = ["llm", "api", "router", "openai", "anthropic", "rate-limit"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries",
]
dependencies = [
    "openai>=1.0.0",
    "pyyaml>=6.0",
    "requests>=2.28.0",
]

[project.optional-dependencies]
anthropic = ["anthropic>=0.20.0"]
dev = ["pytest>=7.0", "pytest-mock"]
all = ["anthropic>=0.20.0", "pytest>=7.0", "pytest-mock"]

[project.urls]
Homepage = "https://github.com/donjguido/llm-router"
Repository = "https://github.com/donjguido/llm-router"
Issues = "https://github.com/donjguido/llm-router/issues"

[tool.setuptools.packages.find]
include = ["llm_router*"]

[tool.setuptools.package-data]
"*" = ["../config/*.yml"]

[tool.pytest.ini_options]
testpaths = ["tests"]
